remove circular imports in lib
